requirements.txt
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
prettytable==3.8.0
schedule==1.2.0
asciichartpy==1.5.25
gunicorn==21.2.0